+++ SMT_QoLity/SuperMarket/PatchClassHelpers/TargetMarking/ShelfSlotInfo/ProductShelfSlotInfo.cs
@@ -1,22 +1,19 @@
 ﻿using UnityEngine;
 
-namespace SuperQoLity.SuperMarket.PatchClassHelpers.TargetMarking.SlotInfo {
+namespace SuperQoLity.SuperMarket.PatchClassHelpers.TargetMarking.ShelfSlotInfo {
 
-	public class ProductShelfSlotInfo : SlotInfoBase {
+	public class ProductShelfSlotInfo : GenericShelfSlotInfo {
 
 
 		public ProductShelfSlotInfo(int shelfIndex, int slotIndex, int productId, int quantity, Vector3 position)
-			: base(shelfIndex, slotIndex, productId, quantity, position) { }
+			: base(shelfIndex, slotIndex, productId, quantity, position, ShelfType.ProdShelfSlot) { }
 
 
 		public ProductShelfSlotInfo(int shelfIndex, int slotIndex)
-			: base(shelfIndex, slotIndex) { }
+			: base(shelfIndex, slotIndex, ShelfType.ProdShelfSlot) { }
 
 
 		public static ProductShelfSlotInfo Default { get { return new ProductShelfSlotInfo(-1, -1, -1, -1, Vector3.zero); } }
-
-
-		public bool FreeProductShelfFound { get { return ShelfIndex >= 0 && SlotIndex >= 0; } }
 
 	}